removed the gameController.js file and split the purpose of the adminGameRoutes.js and the pulicGameRoutes.js file to take over that to make it more clear
